Dashboard/requirements.txt
streamlit
python-dotenv
psycopg2-binary